requirements.txt
absl-py==1.3.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
black==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dill==0.3.6
etils==0.9.0
exceptiongroup==1.0.0rc9
firebase-admin==5.4.0
flatbuffers==22.10.26
fonttools==4.37.4
gast==0.4.0
get-project-root==0.2
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
grpcio-tools==1.44.0
h11==0.14.0
h5py==3.7.0
httplib2==0.20.4
idna==3.4
importlib-resources==5.10.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pathspec==0.10.2
Pillow==9.2.0
platformdirs==2.5.4
promise==2.3
proto-plus==1.22.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-macos==2.10.0
tensorflow-metadata==1.11.0
tensorflow-metal==0.6.0
termcolor==2.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.10.0